+++ tests/html/MultipleTest.php
@@ -33,8 +33,14 @@ final class MultipleTest extends \PHPUnit\Framework\TestCase
 
         $renderer = new \DBlackborough\Quill\Renderer\Html();
 
-        $this->assertEquals($renderer->load($parser->deltasByIndex('one'))->render(), $this->expected_one);
-        $this->assertEquals($renderer->load($parser->deltasByIndex('two'))->render(), $this->expected_two);
+        $this->assertEquals(
+            $this->expected_one,
+            trim($renderer->load($parser->deltasByIndex('one'))->render())
+        );
+        $this->assertEquals(
+            $this->expected_two,
+            trim($renderer->load($parser->deltasByIndex('two'))->render())
+        );
     }
 
     /**
@@ -60,8 +66,52 @@ final class MultipleTest extends \PHPUnit\Framework\TestCase
             $this->fail(__METHOD__ . 'failure, ' . $e->getMessage());
         }
 
-        $this->assertEquals($this->expected_one, $result_one, __METHOD__ . ' multiple text first index failure');
-        $this->assertEquals($this->expected_two, $result_two, __METHOD__ . ' multiple text second index failure');
+        $this->assertEquals(
+            $this->expected_one,
+            trim($result_one),
+            __METHOD__ . ' multiple text first index failure'
+        );
+        $this->assertEquals(
+            $this->expected_two,
+            trim($result_two),
+            __METHOD__ . ' multiple text second index failure'
+        );
+    }
+
+    /**
+     * Parse multiple $quill_json strings and return via index, call via API,
+     * trim option set
+     */
+    public function testLoadingMultipleDeltasAndRenderingWithTrim()
+    {
+        $quill_json = [
+            'one' => $this->delta_one,
+            'two' => $this->delta_two
+        ];
+
+        $result_one = null;
+        $result_two = null;
+
+        try {
+            $quill = new RenderMultiple($quill_json, 'HTML');
+
+            $result_one = $quill->render('one', true);
+            $result_two = $quill->render('two', true);
+
+        } catch (\Exception $e) {
+            $this->fail(__METHOD__ . 'failure, ' . $e->getMessage());
+        }
+
+        $this->assertEquals(
+            $this->expected_one,
+            $result_one,
+            __METHOD__ . ' multiple text first index failure'
+        );
+        $this->assertEquals(
+            $this->expected_two,
+            $result_two,
+            __METHOD__ . ' multiple text second index failure'
+        );
     }
 
     /**