+++ tests/html/StockTest.php
@@ -5,6 +5,7 @@ namespace DBlackborough\Quill\Tests\Html;
 require __DIR__ . '../../../vendor/autoload.php';
 
 use DBlackborough\Quill\Render as QuillRender;
+use DBlackborough\Quill\RenderMultiple as QuillRenderMultiple;
 
 /**
  * Stock tests, testing base functionality not necessarily related to attributes
@@ -35,7 +36,11 @@ final class StockTest extends \PHPUnit\Framework\TestCase
             $this->fail(__METHOD__ . 'failure, ' . $e->getMessage());
         }
 
-        $this->assertEquals($this->expected_null_insert, $result, __METHOD__ . ' Null insert skipped failure');
+        $this->assertEquals(
+            $this->expected_null_insert,
+            trim($result),
+            __METHOD__ . ' Null insert skipped failure'
+        );
     }
 
     /**
@@ -52,15 +57,12 @@ final class StockTest extends \PHPUnit\Framework\TestCase
         $parser = new \DBlackborough\Quill\Parser\Html();
 
         try {
-
-            $parser->load($this->delta_header);
-            $parser->parse();
+            $parser->load($this->delta_header)->parse();
 
             $renderer = new \DBlackborough\Quill\Renderer\Html();
             $result = $renderer->load($parser->deltas())->render();
 
-            $parser->load($this->delta_header);
-            $parser->parse();
+            $parser->load($this->delta_header)->parse();
 
             $renderer = new \DBlackborough\Quill\Renderer\Html();
             $result = $renderer->load($parser->deltas())->render();
@@ -68,7 +70,11 @@ final class StockTest extends \PHPUnit\Framework\TestCase
             $this->fail(__METHOD__ . 'failure, ' . $e->getMessage());
         }
 
-        $this->assertEquals($this->expected_header, $result, __METHOD__ . ' Multiple load calls failure');
+        $this->assertEquals(
+            $this->expected_header,
+            trim($result),
+            __METHOD__ . ' Multiple load calls failure'
+        );
     }
 
     /**
@@ -81,7 +87,20 @@ final class StockTest extends \PHPUnit\Framework\TestCase
     {
         $this->expectException(\InvalidArgumentException::class);
 
-        $quill = new QuillRender($this->delta_header, 'UNKNOWN');
+        new QuillRender($this->delta_header, 'UNKNOWN');
+    }
+
+    /**
+     * Test to see if an exception is thrown when an invalid parser is requested
+     *
+     * @return void
+     * @throws \Exception
+     */
+    public function testExceptionThrownForInvalidMultipleParser()
+    {
+        $this->expectException(\InvalidArgumentException::class);
+
+        new QuillRenderMultiple([], 'UNKNOWN');
     }
 
     /**
@@ -92,8 +111,50 @@ final class StockTest extends \PHPUnit\Framework\TestCase
      */
     public function testExceptionThrownForInvalidJson()
     {
-        $this->expectException(\RuntimeException::class);
+        $this->expectException(\InvalidArgumentException::class);
 
-        $quill = new QuillRender($this->delta_header_invalid, 'HTML');
+        new QuillRender($this->delta_header_invalid, 'HTML');
+    }
+
+    /**
+     * Test to see if an exception is thrown when attempting to parse an invalid json string
+     *
+     * @return void
+     * @throws \Exception
+     */
+    public function testExceptionThrownForInvalidMultipleJson()
+    {
+        $this->expectException(\InvalidArgumentException::class);
+
+        new QuillRenderMultiple(
+            [
+                'delta_1' => $this->delta_header,
+                'delta_2' => $this->delta_header_invalid
+            ],
+            'HTML'
+        );
+    }
+
+    /**
+     * Test the trim option of render
+     *
+     * @return void
+     */
+    public function testTrimOptionOfRender()
+    {
+        $result = null;
+
+        try {
+            $quill = new QuillRender($this->delta_header);
+            $result = $quill->render(true);
+        } catch (\Exception $e) {
+            $this->fail(__METHOD__ . 'failure, ' . $e->getMessage());
+        }
+
+        $this->assertEquals(
+            $this->expected_header,
+            $result,
+            __METHOD__ . ' Trim option failure'
+        );
     }
 }